+++ nusamai/src/sink/minecraft/mod.rs
@@ -1,5 +1,8 @@
 //! Minecraft sink
-
+mod region_writer;
+use region_writer::{
+    write_region, BlockSchema, ChunkSchema, PositionXZ, RegionSchema, SectionSchema, WorldSchema,
+};
 use std::{fs::File, path::PathBuf};
 
 use hashbrown::HashMap;
@@ -11,17 +14,13 @@ use crate::{
     parameters::*,
     pipeline::{Feedback, PipelineError, Receiver, Result},
     sink::{DataRequirements, DataSink, DataSinkProvider, SinkInfo},
+    transformer,
 };
 
 use earcut::{utils3d::project3d_to_2d, Earcut};
 use nusamai_projection::etmerc::ExtendedTransverseMercatorProjection;
 
 use nusamai_voxelize::{DdaVoxelizer, MeshVoxelizer, Voxel};
-
-use fastanvil::Region;
-use fastnbt::{to_bytes, LongArray};
-use serde::{Deserialize, Serialize};
-use std::sync::{Arc, Mutex};
 
 use nusamai_citygml::object::{Map, Object};
 
@@ -102,96 +101,15 @@ use block_colors::get_block_colors;
 mod typename_colors;
 use typename_colors::get_typename_colors;
 
-type PositionXZ = [i32; 2];
-#[derive(Deserialize, Serialize, Debug)]
-struct PositionXYZ([u8; 3]);
-
-impl PositionXYZ {
-    // Check that the input is in the range 0~15
-    fn new(x: u8, y: u8, z: u8) -> Result<Self> {
-        if x > 15 || y > 15 || z > 15 {
-            Err(PipelineError::Canceled)
-        } else {
-            Ok(PositionXYZ([x, y, z]))
-        }
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-struct BlockSchema {
-    position: PositionXYZ,
-    name: String,
-}
-
-impl BlockSchema {
-    fn new(x: u8, y: u8, z: u8, name: String) -> Result<Self> {
-        let position = PositionXYZ::new(x, y, z)?;
-        Ok(BlockSchema { position, name })
-    }
-}
-#[derive(Deserialize, Serialize, Debug)]
-struct SectionSchema {
-    y: i32,
-    blocks: Vec<BlockSchema>,
-}
-#[derive(Deserialize, Serialize, Debug)]
-struct ChunkSchema {
-    position: PositionXZ,
-    sections: Vec<SectionSchema>,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-struct RegionSchema {
-    position: PositionXZ,
-    chunks: Vec<ChunkSchema>,
-}
-
-type WorldSchema = Vec<RegionSchema>;
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Debug)]
-struct Chunk {
-    Status: String, // The status of the chunk, such as whether it is fully generated or being generated.
-    zPos: i32,      // The Z coordinate of the chunk (absolute value).
-    yPos: i32,      // The Y coordinate of the lowest section in the chunk.
-    xPos: i32,      // The X coordinate of the chunk (absolute value).
-    sections: Vec<Section>, // A vector containing the sections that make up the chunk.
-    DataVersion: u32, // The version of the data format used to store this chunk.
-}
-
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Clone, Debug)]
-struct Section {
-    block_states: Blockstates,
-    biomes: Biomes,
-    Y: i8,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-struct Biomes {
-    palette: Vec<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-struct Blockstates {
-    palette: Vec<PaletteItem>,
-    data: Option<LongArray>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct PaletteItem {
-    #[serde(rename = "Name")]
-    name: String,
-    #[serde(rename = "Properties")]
-    properties: Option<serde_json::Value>,
-}
-
 impl DataSink for MinecraftSink {
     fn make_requirements(&self) -> DataRequirements {
         DataRequirements {
-            // use_appearance: true,
-            // resolve_appearance: true,
-            // key_value: crate::transformer::KeyValueSpec::JsonifyObjects,
+            // key_value: transformer::KeyValueSpec::DotNotation,
+            // lod_filter: transformer::LodFilterSpec {
+            //     mode: transformer::LodFilterMode::Lowest,
+            //     ..Default::default()
+            // },
+            // geom_stats: transformer::GeometryStatsSpec::MinMaxHeights,
             ..Default::default()
         }
     }
@@ -203,6 +121,8 @@ impl DataSink for MinecraftSink {
         let mut region_map: HashMap<PositionXZ, usize> = HashMap::new();
         let mut chunk_map: HashMap<(PositionXZ, PositionXZ), usize> = HashMap::new();
         let mut section_map: HashMap<(PositionXZ, PositionXZ, i32), usize> = HashMap::new();
+
+        let block_colors = get_block_colors();
 
         let mut local_bvol = BoundingVolume::default();
         let parcels: Vec<crate::pipeline::Parcel> = upstream.into_iter().collect();
@@ -241,16 +161,14 @@ impl DataSink for MinecraftSink {
                             Value::Object(obj) => {
                                 let typename = &obj.typename.as_ref();
 
-                                println!("Typename: {}", typename);
-
                                 match typename_colors.get(typename) {
                                     Some(color) => {
                                         rgb = *color;
                                     }
-                                    _ => println!("Typename not found"),
+                                    _ => println!("No color found for typename '{}'", typename),
                                 }
                             }
-                            _ => println!("Typename not found"),
+                            _ => println!("The root value is not an object"),
                         }
 
                         let geom_store = entity.geometry_store.read().unwrap();
@@ -267,7 +185,6 @@ impl DataSink for MinecraftSink {
                         // Calculation of centre coordinates
                         let center_lng = (local_bvol.min_lng + local_bvol.max_lng) / 2.0;
                         let center_lat = (local_bvol.min_lat + local_bvol.max_lat) / 2.0;
-                        let min_height = local_bvol.min_height;
 
                         let projection = ExtendedTransverseMercatorProjection::new(
                             center_lng,
@@ -280,7 +197,8 @@ impl DataSink for MinecraftSink {
                             .vertices
                             .iter()
                             .map(|v| match projection.project_forward(v[0], v[1], v[2]) {
-                                Ok((x, y, z)) => [x, z - min_height - 64 as f64, y * -1 as f64],
+                                // To match the Minecraft coordinate system, the y-coordinate is multiplied by -1 and replaced with z
+                                Ok((x, y, z)) => [x, z, y * -1.0],
                                 Err(e) => {
                                     println!("conversion error: {:?}", e);
                                     [f64::NAN, f64::NAN, f64::NAN]
@@ -331,8 +249,6 @@ impl DataSink for MinecraftSink {
                     })
             },
             || {
-                let block_colors = get_block_colors();
-
                 receiver.into_iter().for_each(|feature| {
                     feature.iter().for_each(|(key, voxel)| {
                         let mut block_name = "minecraft:white_wool";
@@ -419,48 +335,7 @@ impl DataSink for MinecraftSink {
                 let _ = world_data.iter().try_for_each(|region| -> Result<()> {
                     feedback.ensure_not_canceled()?;
 
-                    let out_path = PathBuf::from(format!(
-                        "{}/r.{}.{}.mca",
-                        file_path.display(),
-                        region.position[0],
-                        region.position[1]
-                    ));
-
-                    let out_file = File::options()
-                        .read(true)
-                        .write(true)
-                        .truncate(true)
-                        .create(true)
-                        .open(out_path)
-                        .unwrap();
-
-                    let new_region = Arc::new(Mutex::new(Region::new(out_file).unwrap()));
-
-                    (0..32).into_par_iter().for_each(|chunk_z| {
-                        (0..32).into_par_iter().for_each(|chunk_x| {
-                            // Calculate absolute coordinates of chunks
-                            let absolute_chunk_x = region.position[0] * 32 + chunk_x;
-                            let absolute_chunk_z = region.position[1] * 32 + chunk_z;
-
-                            let chunk_data = region
-                                .chunks
-                                .iter()
-                                .find(|c| c.position == [absolute_chunk_x, absolute_chunk_z]);
-
-                            let chunk = create_chunk_structure(
-                                absolute_chunk_x,
-                                absolute_chunk_z,
-                                chunk_data,
-                            );
-
-                            let ser = to_bytes(&chunk).unwrap();
-
-                            let mut region = new_region.lock().unwrap();
-                            region
-                                .write_chunk(chunk_x as usize, chunk_z as usize, &ser)
-                                .unwrap();
-                        });
-                    });
+                    write_region(region, &file_path)?;
 
                     Ok(())
                 });
@@ -479,156 +354,5 @@ impl DataSink for MinecraftSink {
         }
 
         Ok(())
-    }
-}
-
-// Function to calculate the number of bits and data size of a palette.
-const fn calculate_bits_and_size(palette_len: usize) -> (u32, u32) {
-    let bits_per_block = match palette_len {
-        0..=16 => 4,
-        17..=2048 => (palette_len - 1).ilog2() + 1,
-        2049.. => 12,
-    };
-    let data_size = (4096 * bits_per_block + 63) / 64;
-    (bits_per_block, data_size)
-}
-// Functions to create sections
-fn create_chunk_section(
-    blocks: &[BlockSchema],
-    palette: &mut Vec<PaletteItem>,
-    section_y: i32,
-) -> Section {
-    // Calculate the number of bits and data size based on the size of the palette
-    let (bits_per_block, data_size) = calculate_bits_and_size(palette.len());
-
-    // Create a 1D array of 4096 elements and embed the PALETTE index in each block.
-    let mut block_indices = vec![0; 4096];
-    for block in blocks {
-        let PositionXYZ([x, y, z]) = block.position;
-
-        // Calculate the index of the 1D array and store the index of the palette
-        let index = (y as usize) * 256 + (z as usize) * 16 + (x as usize);
-        let palette_index = palette
-            .iter()
-            .position(|b| b.name == block.name)
-            .unwrap_or_else(|| {
-                palette.push(PaletteItem {
-                    name: block.name.clone(),
-                    properties: None,
-                });
-                palette.len() - 1
-            });
-        block_indices[index] = palette_index;
-    }
-
-    // Calculate the number of blocks stored in entry (i64) by BPE
-    let blocks_per_entry = 64 / bits_per_block;
-
-    // Divide 1D array by number of blocks
-    let block_entries: Vec<&[usize]> = block_indices.chunks(blocks_per_entry as usize).collect();
-
-    // Divide 1D arrays by the number of blocks.
-    let mut data = Vec::with_capacity(data_size as usize);
-    for entry in block_entries {
-        let mut value: i64 = 0;
-        for (i, &index) in entry.iter().enumerate() {
-            value |= (index as i64) << (i * bits_per_block as usize);
-        }
-        data.push(value);
-    }
-
-    // Additional padding as required
-    if data_size as usize > data.len() {
-        let padding_size = data_size as usize - data.len();
-        data.extend(std::iter::repeat(0).take(padding_size));
-    }
-
-    Section {
-        block_states: Blockstates {
-            palette: palette.clone(),
-            data: Some(LongArray::new(data)),
-        },
-        biomes: Biomes {
-            palette: vec!["minecraft:the_void".to_string()],
-        },
-        Y: section_y as i8,
-    }
-}
-
-// Functions to create chunk structures
-fn create_chunk_structure(chunk_x: i32, chunk_z: i32, chunk_data: Option<&ChunkSchema>) -> Chunk {
-    let palette = vec![PaletteItem {
-        name: "minecraft:air".to_string(),
-        properties: None,
-    }];
-
-    let sections: Vec<Section> = if let Some(chunk_data) = chunk_data {
-        chunk_data
-            .sections
-            .iter()
-            .map(|section| {
-                let mut local_palette = palette.clone();
-
-                // Register the block in the palette.
-                for block in &section.blocks {
-                    if !local_palette.iter().any(|b| b.name == block.name) {
-                        local_palette.push(PaletteItem {
-                            name: block.name.clone(),
-                            properties: if block.name == "minecraft:oak_leaves" {
-                                Some(serde_json::json!({ "persistent": "true" }))
-                            } else {
-                                None
-                            },
-                        });
-                    }
-                }
-
-                create_chunk_section(&section.blocks, &mut local_palette, section.y)
-            })
-            .collect()
-    } else {
-        // Create an empty section if there is no chunk_data
-        Vec::new()
-    };
-
-    Chunk {
-        Status: "full".to_string(),
-        zPos: chunk_z,
-        yPos: -4, // Lowest Y section position in the chunk (e.g., -4 in version 1.18 and later)
-        xPos: chunk_x,
-        sections,
-        DataVersion: 3105, // Java Edition 1.19
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_calculate_bits_and_size() {
-        let test_cases = vec![
-            (1, (4, 256)),
-            (16, (4, 256)),
-            (17, (5, 320)),
-            (32, (5, 320)),
-            (33, (6, 384)),
-            (64, (6, 384)),
-            (65, (7, 448)),
-            (128, (7, 448)),
-            (129, (8, 512)),
-            (256, (8, 512)),
-            (257, (9, 576)),
-            (512, (9, 576)),
-            (513, (10, 640)),
-            (1024, (10, 640)),
-            (1025, (11, 704)),
-            (2048, (11, 704)),
-            (2049, (12, 768)),
-        ];
-
-        for (palette_len, expected) in test_cases {
-            assert_eq!(calculate_bits_and_size(palette_len), expected);
-        }
     }
 }
